💄 Design : 사용자, 사업자 마이페이지 index.tsx 파일 수정 #6 #9
